Change return from Float to Double

+++ MathUtilites/Interpolation/Liniar.cs
@@ -10,8 +10,8 @@
         public static double Interpolate(double leftValue, double rightValue, double step)
             => (double)Interpolate((decimal)leftValue, (decimal)rightValue, (decimal)step);
 
-        public static float Interpolate(float leftValue, float rightValue, float step)
-           => (float)Interpolate((decimal)leftValue, (decimal)rightValue, (decimal)step);
+        public static double Interpolate(float leftValue, float rightValue, float step)
+           => (double)Interpolate((decimal)leftValue, (decimal)rightValue, (decimal)step);
 
         public static double Interpolate(int leftValue, int rightValue, int step)
            => (double)Interpolate(leftValue, rightValue, (decimal)step);
